Переделываю features items для home/Index: ViewComponent чтобы из Index не передовать данные)
