+++ Tests/WebStore.Tests/Controllers/HomeControllerTests.cs
@@ -1,0 +1,22 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using WebStore.Controllers;
+using Assert =  Xunit.Assert;
+
+namespace WebStore.Tests.Controllers
+{
+    [TestClass]
+    public class HomeControllerTests
+    {
+        [TestMethod]
+        public void Index_Returns_View()
+        {
+            var controller = new HomeController();
+
+            var result = controller.Index();
+
+            //Assert.IsInstanceOfType(result, typeof(ViewResult));
+            Assert.IsType<ViewResult>(result);
+        }
+    }
+}
