+++ UI/WebStore/Components/SectionsViewComponent.cs
@@ -1,11 +1,8 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
 using WebStore.Interfaces.Services;
 using WebStore.Domain.ViewModels;
-using WebStore.Services.Mapping;
 
 namespace WebStore.Components
 {
