+++ UI/WebStore/Views/Shared/Components/features_items/Default.cshtml
@@ -1,0 +1,58 @@
+﻿@model IEnumerable<WebStore.Domain.ViewModels.ProductViewModel>
+
+<div class="features_items">
+    <!--features_items-->
+    <h2 class="title text-center">Товары</h2>
+    @if (Model is null)
+    {
+        <p>Товаров нет(</p>
+    }
+    else
+    {
+        @foreach (var product in Model)
+        {
+            <div class="col-sm-4">
+                <div class="product-image-wrapper">
+                    <div class="single-products">
+                        <div class="productinfo text-center">
+                            <img src="~/images/shop/@product.ImageUrl" alt="" />
+                            <h2>@product.Price.ToString("C")</h2>
+                            <p>
+                                <a asp-controller="Catalog" asp-action="Details" asp-route-id="@product.Id">@product.Name</a>
+                            </p>
+                            <a asp-controller="Cart" asp-action="Add" asp-route-id="@product.Id"
+                               class="btn btn-default add-to-cart">
+                                <i class="fa fa-shopping-cart"></i>В корзину
+                            </a>
+                        </div>
+                        <div class="product-overlay">
+                            <div class="overlay-content">
+                                <h2>@product.Price.ToString("C")</h2>
+                                <p>
+                                    <a asp-controller="Catalog" asp-action="Details" asp-route-id="@product.Id">@product.Name</a>
+                                </p>
+                                <a asp-controller="Cart" asp-action="Add" asp-route-id="@product.Id"
+                                   class="btn btn-default add-to-cart">
+                                    <i class="fa fa-shopping-cart"></i>В корзину
+                                </a>
+                            </div>
+                        </div>
+                    </div>
+                    <div class="choose">
+                        <ul class="nav nav-pills nav-justified">
+                            <li><a href=""><i class="fa fa-plus-square"></i>Добавить в желания</a></li>
+                            <li><a href=""><i class="fa fa-plus-square"></i>Добавить к сравнению</a></li>
+                        </ul>
+                    </div>
+                </div>
+            </div>
+        }
+    }
+
+    <ul class="pagination">
+        <li class="active"><a href="">1</a></li>
+        <li><a href="">2</a></li>
+        <li><a href="">3</a></li>
+        <li><a href="">&raquo;</a></li>
+    </ul>
+</div><!--features_items-->
